# openapi.yaml
openapi: 3.1.0
info:
  title: Qarax API
  description: REST API for managing virtual machines and hypervisor hosts
  license:
    name: ''
  version: 0.1.0
paths:
...
  /security-groups:
    get:
      tags:
      - security-groups
      operationId: list
      parameters:
      - name: name
        in: query
        description: Optional name filter for list queries
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '200':
          description: List security groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityGroup'
        '500':
          description: Internal server error
    post:
      tags:
      - security-groups
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSecurityGroup'
        required: true
      responses:
        '201':
          description: Security group created successfully
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /security-groups/{security_group_id}:
    get:
      tags:
      - security-groups
      operationId: get
      parameters:
      - name: security_group_id
        in: path
        description: Security group unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Security group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityGroup'
        '404':
          description: Security group not found
        '500':
          description: Internal server error
    delete:
      tags:
      - security-groups
      operationId: delete
      parameters:
      - name: security_group_id
        in: path
        description: Security group unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Security group deleted successfully
        '404':
          description: Security group not found
        '500':
          description: Internal server error
  /security-groups/{security_group_id}/rules:
    get:
      tags:
      - security-groups
      operationId: list_rules
      parameters:
      - name: security_group_id
        in: path
        description: Security group unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List security-group rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityGroupRule'
        '404':
          description: Security group not found
        '500':
          description: Internal server error
    post:
      tags:
      - security-groups
      operationId: create_rule
      parameters:
      - name: security_group_id
        in: path
        description: Security group unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSecurityGroupRule'
        required: true
      responses:
        '201':
          description: Security-group rule created successfully
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Security group not found
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /security-groups/{security_group_id}/rules/{rule_id}:
    delete:
      tags:
      - security-groups
      operationId: delete_rule
      parameters:
      - name: security_group_id
        in: path
        description: Security group unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: rule_id
        in: path
        description: Security-group rule unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Security-group rule deleted successfully
        '404':
          description: Security group or rule not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    NewSecurityGroup:
      type: object
      required:
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        name:
          type: string
    NewSecurityGroupRule:
      type: object
      required:
      - direction
      - protocol
      properties:
        cidr:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        direction:
          $ref: '#/components/schemas/SecurityGroupDirection'
        port_end:
          type:
          - integer
          - 'null'
          format: int32
        port_start:
          type:
          - integer
          - 'null'
          format: int32
        protocol:
          $ref: '#/components/schemas/SecurityGroupProtocol'
...
    SecurityGroup:
      type: object
      required:
      - id
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        name:
          type: string
    SecurityGroupDirection:
      type: string
      enum:
      - ingress
      - egress
    SecurityGroupProtocol:
      type: string
      enum:
      - any
      - tcp
      - udp
      - icmp
    SecurityGroupRule:
      type: object
      required:
      - id
      - security_group_id
      - direction
      - protocol
      properties:
        cidr:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        direction:
          $ref: '#/components/schemas/SecurityGroupDirection'
        id:
          type: string
          format: uuid
        port_end:
          type:
          - integer
          - 'null'
          format: int32
        port_start:
          type:
          - integer
          - 'null'
          format: int32
        protocol:
          $ref: '#/components/schemas/SecurityGroupProtocol'
        security_group_id:
          type: string
          format: uuid
...